pyproject.toml
[build-system]
requires = ["setuptools>=82.0.1", "setuptools-scm>=10.0.5"]
build-backend = "setuptools.build_meta"

[project]
name = "Red-GitHubBot"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.13.3",
    "apscheduler[sqlalchemy]>=3.11.2",
    "cachetools>=7.0.5",
    "cherry-picker>=2.4.0",
    "discord.py>=2.7.1",
    "gidgethub>=5.4.0",
    "graphql-builder==1.0.0a3",
    "markdownify>=1.2.2",
    "mistune>=3",
    "psycopg2>=2.9.11",
    "regex>=2026.2.28",
    "sentry-sdk[pure-eval]>=2.58.0",
    "yarl>=1.23.0",
]
dynamic = ["version"]

[tool.setuptools_scm]

[dependency-groups]
dev = [
    "ruff>=0.15.11",
]

[tool.uv]
cache-keys = [{ file = "pyproject.toml" }, { git = { commit = true } }]

[tool.ruff]
line-length = 99

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = [
    # mccabe
    "C90",
    # pyflakes rules
    "F",
    # pycodestyle rules
    "E",
    "W",
    # import sorting
    "I",
    # pygrep-hooks
    "PGH",
]
ignore = [
    "E203",
]

[tool.ruff.lint.isort]
combine-as-imports = true
